+++ lib/src/entities/featured_tag.dart
@@ -19,7 +19,7 @@ class FeaturedTag with _$FeaturedTag {
     /// The number of authored statuses containing this hashtag.
     @JsonKey(name: 'statuses_count') final String? statusesCount,
 
-    /// The timestamp of the last authored status containing this hashtag.
+    /// The date of the last authored status containing this hashtag.
     @JsonKey(name: 'last_status_at') final DateTime? lastStatusAt,
   }) = _FeaturedTag;